Code refactor with getHandlers.
